Add to cart (#59)
* This add @peAuthorize

* this is used to test the cloudinary sever

* this change modify the add to favourites endpoint

* This modify the swagger configuration

* this modify add to product to favourites class

* this modify the cart class

* this remove cart class from the customerOrder class

* this modify orderDetails class

* this add test to the updateserviceimpl

* modified files

* Change non private access to private access

* resolved conflict

+++ src/main/java/com/example/safariwebstore008/models/Product.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Null;
+import java.math.BigInteger;
 import java.util.List;
 
 
@@ -25,7 +26,7 @@ public class Product extends BaseClass {
     private String description;
 
     @NotNull(message = "product price field is empty")
-    private Double price;
+    private BigInteger price;
 
     private String colors;
